Variable Stroke Width
Add the option to either set a fixed width for all lines (as always) or individually per series as an array of widths.

Some marshalling tricks were needed as the `width` property can either be an array or a number.

If both set, the width has preference over the array.

Bump version to 23.0.2